+++ mongodb_knowledge_graph/storage/mongodb_adapter.py
@@ -22,8 +22,8 @@ except ImportError as e:
         "PyMongo is required for MongoDB storage. Install with: pip install pymongo"
     ) from e
 
-from mcp_server_memory.storage.base import StorageAdapter, StorageError
-from mcp_server_memory.types import Entity, KnowledgeGraph, Relation
+from mongodb_knowledge_graph.storage.base import StorageAdapter, StorageError
+from mongodb_knowledge_graph.types import Entity, KnowledgeGraph, Relation
 
 
 class MongoDBStorageAdapter(StorageAdapter):
